+++ App/Source/ViewController/기타/CoordinatesConversionViewController.swift
@@ -1,11 +1,3 @@
-//
-//  CoordinatesConversionViewController.swift
-//  iNaviMapsDemo
-//
-//  Created by 장은석 on 2022/08/11.
-//  Copyright © 2022 DAECHEOL KIM. All rights reserved.
-//
-
 import UIKit
 import iNaviMaps